+++ src/query/range_query/mod.rs
@@ -1,0 +1,5 @@
+mod fast_field_range_query;
+mod range_query;
+mod range_query_ip_fastfield;
+
+pub use self::range_query::RangeQuery;
